adjusted error states

+++ data/src/main/java/br/com/andretortolano/data/sources/local/FakeLocalSource.kt
@@ -3,7 +3,7 @@ package br.com.andretortolano.data.sources.local
 import br.com.andretortolano.data.models.NumberTriviaModel
 
 class FakeLocalSource: LocalSource {
-    override fun getNumberTrivia(number: Int): NumberTriviaModel? {
+    override fun getNumberTrivia(number: Long): NumberTriviaModel? {
         return null
     }
