+++ data/src/main/java/br/com/andretortolano/data/sources/local/LocalSource.kt
@@ -3,7 +3,7 @@ package br.com.andretortolano.data.sources.local
 import br.com.andretortolano.data.models.NumberTriviaModel
 
 interface LocalSource {
-    fun getNumberTrivia(number: Int): NumberTriviaModel?
+    fun getNumberTrivia(number: Long): NumberTriviaModel?
 
     fun saveNumberTrivia(expectedNumberTrivia: NumberTriviaModel)
 }
