+++ data/src/main/java/br/com/andretortolano/data/sources/remote/RemoteSource.kt
@@ -3,7 +3,7 @@ package br.com.andretortolano.data.sources.remote
 import br.com.andretortolano.data.models.NumberTriviaModel
 
 interface RemoteSource {
-    fun getConcreteNumberTrivia(number: Int): NumberTriviaModel
+    fun getConcreteNumberTrivia(number: Long): NumberTriviaModel
 
     fun getRandomNumberTrivia(): NumberTriviaModel
 }
