+++ data/src/main/java/br/com/andretortolano/data/sources/remote/retrofit/NumberTriviaJson.kt
@@ -1,10 +1,8 @@
 package br.com.andretortolano.data.sources.remote.retrofit
 
-import java.math.BigInteger
-
 data class NumberTriviaJson(
     val text: String,
-    val number: BigInteger,
+    val number: Long,
     val found: Boolean,
     val type: String
 )
